Move activation functions to flax_utils.
